Use "real" Kubernetes deployment manifest
The way in which the kubedex-exporter is being scraped by Prometheus has
also been changed. In theory it should be possible to scale the
kubedex-exporter horizontally (it makes no sense, probably,...) and
scraping will happen through a Service that sits in front of the
deployed ReplicaSet, *if* `service.enabled` is set to `true` during
deployment of the Chart.

This fixes #1